requirements.txt
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
